# CommonLib/src/main/resources/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "openstockæ��ä¾›APIè°ƒç”¨ï¼ŒæŸ¥è¯¢è‚¡ç¥¨ä»·æ ¼ï¼ŒæŸ¥è¯¢æŒ‡æ ‡æ•°æ�®ï¼ŒæŸ¥è¯¢æŒ‡æ ‡ç»Ÿè®¡ï¼Œæ¨¡æ‹Ÿè‚¡ç¥¨èµ°åŠ¿ï¼ŒKçº¿å›¾ä¸‹è½½ç­‰ã€‚",
    "version": "1.0.0",
    "title": "OpenStock",
    "termsOfService": "http://192.168.1.12:8080",
    "contact": {
      "email": "éª‘ç‰›ä¸Šå±±"
    },
    "license": {
      "name": "N/A ",
      "url": "N/A"
    }
  },
  "host": "192.168.1.12:8080",
  "basePath": "/portal",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/statistics/luzao/trend/{date}": {
      "get": {
        "tags": [
          "Statistics"
        ],
        "summary": "é²�å…†å¤§è¶‹åŠ¿ç»Ÿè®¡",
        "description": "é²�å…†å¤§è¶‹åŠ¿ç»Ÿè®¡",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "date",
            "description": "æ—¥æœŸåŒºé—´(ä¾‹å¦‚2017-01-03_2017-02-23)",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "æˆ�åŠŸè¿”å›ž"
          }
        }
      }
    },
    "/statistics/luzao/gordon/{date}": {
      "get": {
        "tags": [
          "Statistics"
        ],
        "summary": "é²�å…†å¤§é‡‘å�‰ç»Ÿè®¡",
        "description": "é²�å…†å¤§é‡‘å�‰ç»Ÿè®¡",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "date",
            "description": "æ—¥æœŸåŒºé—´(ä¾‹å¦‚2017-01-03_2017-02-23)",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "æˆ�åŠŸè¿”å›ž"
          }
        }
      }
    },
    "/statistics/qsdd/{date}": {
      "get": {
        "tags": [
          "Statistics"
        ],
        "summary": "è¶‹åŠ¿é¡¶åº•ç»Ÿè®¡",
        "description": "è¶‹åŠ¿é¡¶åº•ç»Ÿè®¡",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "date",
            "description": "æ—¥æœŸåŒºé—´(ä¾‹å¦‚2017-01-03_2017-02-23)",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "æˆ�åŠŸè¿”å›ž"
          }
        }
      }
    },
    "/statistics/shenxian/{date}": {
      "get": {
        "tags": [
          "Statistics"
        ],
        "summary": "ç¥žä»™å¤§è¶‹åŠ¿ç»Ÿè®¡",
        "description": "ç¥žä»™å¤§è¶‹åŠ¿ç»Ÿè®¡",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "date",
            "description": "æ—¥æœŸåŒºé—´(ä¾‹å¦‚2017-01-03_2017-02-23)",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "æˆ�åŠŸè¿”å›ž"
          }
        }
      }
    },
    "/statistics/macd/{date}": {
      "get": {
        "tags": [
          "Statistics"
        ],
        "summary": "MACDç»Ÿè®¡",
        "description": "MACDç»Ÿè®¡",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "date",
            "description": "æ—¥æœŸåŒºé—´(ä¾‹å¦‚2017-01-03_2017-02-23)",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "æˆ�åŠŸè¿”å›ž"
          }
        }
      }
    },
    "/statistics/wr/{date}": {
      "get": {
        "tags": [
          "Statistics"
        ],
        "summary": "å¨�å»‰æŒ‡æ ‡ç»Ÿè®¡",
        "description": "å¨�å»‰æŒ‡æ ‡ç»Ÿè®¡",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "date",
            "description": "æ—¥æœŸåŒºé—´(ä¾‹å¦‚2017-01-03_2017-02-23)",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "æˆ�åŠŸè¿”å›ž"
          }
        }
      }
    }
  },
...
}